FELIX: Add xml option for activating GLP
This commit adds the option "Use GLP" to "FELIX Physical Parameters".
If true, the grounding line parameterization is applied to any basal
friction law that is spatially varying.  The GLP simply evaluates the
floatation criterion at each cubature point, zeroing it if that cubature
point coincides with floating ice.  The basal friction field should be
nonzero everywhere for this to evaluate properly.

The default for "Use GLP" is true.